feat: 고객센터 FAQ 기능 구현

+++ src/main/java/com/example/travlediary/model/Faq.java
@@ -12,7 +12,7 @@ public class Faq {
     private Long id; // 자주묻는질문 번호
     private String question; // 질문
     private String answer; // 답변
-    private Integer orderIndex; // faq 정렬 순서
+    private Long orderIndex; // faq 정렬 순서
     private Boolean isVisible; // 표시여부 숨김, 보여짐 TRUE 노출, FALSE 숨김
     private Timestamp createdAt; // 생성일
     private Timestamp updatedAt; // 수정일
